requirements.txt
altair==4.1.0
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.5
click==7.1.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==2.2.10
django-admin-interface==0.15.5
django-colorfield==0.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.5.0
django-multiforloop==0.2.0
entrypoints==0.3
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
MarkupPy==1.14
MarkupSafe==1.1.1
nltk==3.5
numpy==1.20.1
odfpy==1.4.1
openpyxl==3.0.6
pandas==1.2.3
pycodestyle==2.6.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
rake-nltk==1.0.4
regex==2020.11.13
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
tablib==3.0.0
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tqdm==4.59.0
Unipath==1.1
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
